python-projects/multiclass-classifier-service/requirements.txt
Flask==2.3.2
torch==2.0.1
transformers==4.31.0
python-docx==0.8.11
nltk==3.7
textblob==0.17.1
pandas==1.5.3
sentence-transformers==3.2.0